Check before dropping a foreign key in the users table

+++ database/migrations/2021_10_25_062314_user_delete_foreign_key.php
@@ -14,8 +14,16 @@ class UserDeleteForeignKey extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
+
+            Schema::disableForeignKeyConstraints();
             $table->dropForeign('users_invite_id_foreign_custom');
-            $table->dropColumn('users_invite_id_foreign');
+
+
+            if (Schema::hasColumn('users', 'users_invite_id_foreign')) {
+                $table->dropColumn('users_invite_id_foreign');
+            }
+            Schema::enableForeignKeyConstraints();
+
         });
     }
